improve isolation of versions
