+++ src/app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\LoginRequest;
-use App\Http\Resources\UserPublicInfoResource;
+use App\Http\Resources\UserResource;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
@@ -14,13 +14,13 @@ class AuthenticatedSessionController extends Controller
     /**
      * Handle an incoming authentication request.
      */
-    public function store(LoginRequest $request): UserPublicInfoResource
+    public function store(LoginRequest $request): UserResource
     {
         $request->authenticate();
 
         $request->session()->regenerate();
 
-        return new UserPublicInfoResource($request->user());
+        return new UserResource($request->user());
     }
 
     /**
