+++ src/app/Http/Controllers/UserController.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Http\Resources\UserResource;
+use Illuminate\Http\Request;
+
+class UserController extends Controller
+{
+  /**
+   * Display the specified resource.
+   */
+  public function show(Request $request)
+  {
+    return new UserResource($request->user());
+  }
+
+  /**
+   * Update the specified resource in storage.
+   */
+  public function update(Request $request, string $id)
+  {
+    //
+  }
+
+  /**
+   * Remove the specified resource from storage.
+   */
+  public function destroy(string $id)
+  {
+    //
+  }
+}
