+++ src/config/app.php
@@ -215,4 +215,10 @@ return [
         // 'ExampleClass' => App\Example\ExampleClass::class,
     ])->toArray(),
 
+    /*
+    |--------------------------------------------------------------------------
+    | App Version
+    |--------------------------------------------------------------------------
+    */
+    'version' => env('APP_VERSION'),
 ];
