+++ src/routes/api.php
@@ -1,8 +1,6 @@
 <?php
 
 use App\Helpers\RouteHelper;
-use App\Http\Resources\UserPublicInfoResource;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -16,10 +14,6 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
-    return new UserPublicInfoResource($request->user());
-})->name('user');
-
 Route::prefix('v1')->group(function () {
-    RouteHelper::includeRouteFiles(__DIR__.'/api/v1');
+  RouteHelper::includeRouteFiles(__DIR__.'/api/v1');
 });
