+++ src/routes/api/v1/tier_lists.php
@@ -3,4 +3,4 @@
 use App\Http\Controllers\TierListController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/tierlist/recent', [TierListController::class, 'recent'])->name('tierlist.recent');
+Route::get('/tierlist/recent', [TierListController::class, 'recent'])->name('v1.tierlist.recent');
