+++ src/routes/api/v1/user.php
@@ -1,0 +1,6 @@
+<?php
+
+use App\Http\Controllers\UserController;
+use Illuminate\Support\Facades\Route;
+
+Route::get('/user', [UserController::class, 'show'])->middleware(['auth:sanctum'])->name('v1.user.show');
