--- src/tests/Feature/Auth/AuthenticationTest.php
@@ -1,39 +1,0 @@
-<?php
-
-namespace Tests\Feature\Auth;
-
-use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\Feature\Traits\UserPublicInfoJsonStructure;
-use Tests\TestCase;
-
-class AuthenticationTest extends TestCase
-{
-    use UserPublicInfoJsonStructure;
-    use RefreshDatabase;
-
-    public function test_users_can_authenticate_using_the_login_screen(): void
-    {
-        $user = User::factory()->create();
-
-        $response = $this->postJson('/login', [
-            'email' => $user->email,
-            'password' => 'password',
-        ]);
-
-        $this->assertAuthenticated();
-        $response->assertJsonStructure($this->userPublicInfoJsonStructure());
-    }
-
-    public function test_users_can_not_authenticate_with_invalid_password(): void
-    {
-        $user = User::factory()->create();
-
-        $this->post('/login', [
-            'email' => $user->email,
-            'password' => 'wrong-password',
-        ]);
-
-        $this->assertGuest();
-    }
-}
