--- src/tests/Feature/Traits/UserPublicInfoJsonStructure.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Tests\Feature\Traits;
-
-trait UserPublicInfoJsonStructure
-{
-    private function userPublicInfoJsonStructure()
-    {
-        return [
-            'data' => [
-                'id',
-                'username',
-                'email',
-                'is_admin',
-                'email_verified',
-                'oauth_provider',
-            ],
-        ];
-    }
-}
