--- src/tests/Feature/Traits/UserWithEmailAlreadyExistsJsonStructure.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Tests\Feature\Traits;
-
-use App\Helpers\StatusHelper;
-
-trait UserWithEmailAlreadyExistsJsonStructure
-{
-    private function userWithEmailAlreadyExistsJsonStructure()
-    {
-        return [
-            'message' => StatusHelper::UserWithEmailAlreadyExists,
-        ];
-    }
-}
