+++ src/tests/Feature/v1/Traits/UserWithEmailAlreadyExistsJsonStructure.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\Feature\V1\Traits;
+
+use App\Helpers\StatusHelper;
+
+trait UserWithEmailAlreadyExistsJsonStructure
+{
+  private function userWithEmailAlreadyExistsJsonStructure()
+  {
+    return [
+        'message' => StatusHelper::UserWithEmailAlreadyExists,
+    ];
+  }
+}
