+++ src/tests/Unit/ExampleTest.php
@@ -6,11 +6,11 @@ use PHPUnit\Framework\TestCase;
 
 class ExampleTest extends TestCase
 {
-    /**
-     * A basic test example.
-     */
-    public function test_that_true_is_true(): void
-    {
-        $this->assertTrue(true);
-    }
+  /**
+   * A basic test example.
+   */
+  public function test_that_true_is_true(): void
+  {
+    $this->assertTrue(true);
+  }
 }
